requirements.txt
opencv-python
pytesseract
streamlit
gradio
torch
transformers
pillow
